VLR: Fixed broken SIGINT signal handler so ctrl-c on console finally works... VLR: Fixed broken kill VikingExtractor child process... VLR: Fixed farewell page should set_page_complete to true... VLR: Added broken introspection safe credits gui... VLR: Minor Glade aesthetics... VLR: Removed some unnecessary / old debugging calls... VLR: Updated some copyright notices I forgot about... VikingExtractor: ReadMe

+++ Extras/Viking Lander Remastered/Launcher/Source/RecoveryPage.py
@@ -241,8 +241,9 @@ class RecoveryPageProxy():
             return
 
         # Kill it...
-        os.kill(self.processID, signal.SIGKILL)
-        self.processID = 0
+        if self.processID != 0:
+            os.kill(self.processID, signal.SIGKILL)
+            self.processID = 0
 
     # VikingExtractor terminated...
     def onChildProcessExit(self, *junk):
